requirements.txt
alembic==1.4.2
argon2-cffi==20.1.0
astropy==4.0.1.post1
attrs==19.3.0
azure-core==1.7.0
azure-storage-blob==12.3.2
backcall==0.2.0
bleach==3.1.5
blis==0.4.1
boto==2.49.0
boto3==1.14.36
botocore==1.17.36
catalogue==1.0.0
catboost==0.24
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
colorlover==0.3.0
combo==0.1.1
confuse==1.3.0
cryptography==3.0
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.3
databricks-cli==0.11.0
datefinder==0.7.1
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
docker==4.2.2
docutils==0.15.2
entrypoints==0.3
Flask==1.1.2
funcy==1.14
future==0.18.2
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.7
gorilla==0.3.0
graphviz==0.14.1
gunicorn==20.0.4
htmlmin==0.1.12
idna==2.10
ImageHash==4.1.0
imbalanced-learn==0.7.0
importlib-metadata==1.7.0
importlib-resources==3.0.0
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
kmodes==0.10.2
lightgbm==2.3.1
llvmlite==0.33.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.0
missingno==0.4.2
mistune==0.8.4
mlflow==1.10.0
mlxtend==0.17.3
more-itertools==8.4.0
msrest==0.6.18
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.5
notebook==6.1.1
numba==0.50.1
numexpr==2.7.1
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
pandas==1.1.0
pandas-profiling==2.8.0
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
plotly==4.9.0
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.8.0
prometheus-flask-exporter==0.15.4
prompt-toolkit==3.0.5
protobuf==3.12.4
ptyprocess==0.6.0
py==1.9.0
pycaret==2.0
pycparser==2.20
Pygments==2.6.1
pyLDAvis==2.1.2
pyod==0.8.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
querystring-parser==1.2.4
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
smart-open==2.1.0
smmap==3.0.4
spacy==2.3.2
SQLAlchemy==1.3.13
sqlparse==0.3.1
srsly==1.0.2
suod==0.0.4
sweetviz==1.0b5
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
umap-learn==0.4.6
urllib3==1.25.10
visions==0.4.4
wasabi==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.7.0
xgboost==1.1.1
yellowbrick==1.1
zipp==3.1.0
zope.interface==5.1.0
